@@ -1,34 +1,86 @@
 {
-  "name": "eventnative-ui",
+  "name": "cra-en",
   "version": "0.1.0",
   "private": true,
-  "test": "jest",
   "dependencies": {
     "@ant-design/icons": "^4.2.2",
+    "@babel/core": "7.9.0",
     "@ksense/eventnative": "^1.0.11",
-    "@types/node": "^14.6.4",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
-    "antd": "^4.6.2",
-    "autoprefixer": "^10.0.1",
+    "@svgr/webpack": "4.3.3",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "@types/react-router-dom": "^5.1.6",
+    "@typescript-eslint/eslint-plugin": "^2.10.0",
+    "@typescript-eslint/parser": "^2.10.0",
+    "antd": "^4.7.2",
     "axios": "^0.20.0",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "8.1.0",
+    "babel-plugin-named-asset-import": "^0.3.6",
+    "babel-preset-react-app": "^9.1.2",
     "bizcharts": "^4.0.14",
-    "firebase": "^7.21.0",
+    "camelcase": "^5.3.1",
+    "case-sensitive-paths-webpack-plugin": "2.3.0",
+    "css-loader": "3.4.2",
+    "dotenv": "8.2.0",
+    "dotenv-expand": "5.1.0",
+    "eslint": "^6.6.0",
+    "eslint-config-react-app": "^5.2.1",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-flowtype": "4.6.0",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.19.0",
+    "eslint-plugin-react-hooks": "^1.6.1",
+    "file-loader": "4.3.0",
+    "firebase": "^7.24.0",
+    "fs-extra": "^8.1.0",
     "highlight.run": "^1.0.26",
-    "lodash": "^4.17.20",
-    "logrocket": "^1.0.13",
-    "posthog-js": "^1.4.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-firebaseui": "^4.1.0",
-    "react-router": "^5.2.0",
+    "html-webpack-plugin": "4.0.0-beta.11",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "24.9.0",
+    "jest-environment-jsdom-fourteen": "1.0.1",
+    "jest-resolve": "24.9.0",
+    "jest-watch-typeahead": "0.4.2",
+    "less": "^3.12.2",
+    "less-loader": "^7.0.2",
+    "logrocket": "^1.0.14",
+    "mini-css-extract-plugin": "0.9.0",
+    "optimize-css-assets-webpack-plugin": "5.0.3",
+    "pnp-webpack-plugin": "1.6.4",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-normalize": "8.0.1",
+    "postcss-preset-env": "6.7.0",
+    "postcss-safe-parser": "4.0.1",
+    "posthog-js": "^1.5.0",
+    "react": "^17.0.0",
+    "react-app-polyfill": "^1.0.6",
+    "react-dev-utils": "^10.2.1",
+    "react-dom": "^17.0.0",
     "react-router-dom": "^5.2.0",
-    "react-syntax-highlighter": "^15.1.0"
+    "react-syntax-highlighter": "^15.2.1",
+    "resolve": "1.15.0",
+    "resolve-url-loader": "3.1.2",
+    "sass-loader": "8.0.2",
+    "semver": "6.3.0",
+    "style-loader": "0.23.1",
+    "terser-webpack-plugin": "2.3.8",
+    "ts-pnp": "1.1.6",
+    "typescript": "^4.0.3",
+    "url-loader": "2.3.0",
+    "uuid": "^8.3.1",
+    "webpack": "4.42.0",
+    "webpack-dev-server": "3.11.0",
+    "webpack-manifest-plugin": "2.2.0",
+    "workbox-webpack-plugin": "4.3.1"
   },
   "scripts": {
-    "start": "ENV=dev webpack-dev-server --open --mode development",
-    "build": "ENV=prod webpack --mode production",
-    "analyze": "source-map-explorer ./dist/*.js"
+    "start": "TSC_COMPILE_ON_ERROR=true EXTEND_ESLINT=true node scripts/start.js",
+    "build": "GENERATE_SOURCEMAP=false TSC_COMPILE_ON_ERROR=true EXTEND_ESLINT=true node scripts/build.js",
+    "test": "node scripts/test.js"
   },
   "eslintConfig": {
     "extends": "react-app"
