@@ -1,10 +1,10 @@
 import * as React from 'react'
 import {Button, Card, Col, Form, Input, message, Modal, Row} from "antd";
-import {LockOutlined, UserOutlined, MailOutlined} from "@ant-design/icons/lib";
+import {LockOutlined, UserOutlined, MailOutlined} from "@ant-design/icons";
 
-const logo = require('../../../icons/logo.svg').default;
-const googleLogo = require('../../../icons/google.svg').default;
-const githubLogo = require('../../../icons/github.svg').default;
+const logo = require('../../../icons/logo.svg');
+const googleLogo = require('../../../icons/google.svg');
+const githubLogo = require('../../../icons/github.svg');
 import './LoginForm.less'
 import ApplicationServices from "../../services/ApplicationServices";
 import {navigateAndReload, reloadPage} from "../../commons/utils";
